Move mooncake connector to separate folder
