+++ hs_connectors/src/hs_connectors/mooncake_store.py
